Add Go 1.11 Module Support
As of Go 1.11, go applications can specify their dependencies in a `go.mod` file.
Checksums of these dependencies are stored in a `go.sum` file.

With these files in place, one guarantees reproducable builds of a given binary.

Additionally, this allows go application development to be done outside of the normal `GOPATH`.

Hope this PR finds you well, and happy #HACKTOBER!